Make changing the current minute update the cached states values as well

+++ state/state.go
@@ -2179,8 +2179,8 @@ func (s *State) SetLeaderTimestamp(ts interfaces.Timestamp) {
 	s.LeaderTimestamp = ts //SetLeaderTimestamp()
 	s.LogPrintf("executeMsg", "Set LeaderTimeStamp %d %v for %s", s.LLeaderHeight, s.LeaderTimestamp.String(), atomic.WhereAmIString(1))
 }
-func (s *State) SetLLeaderHeight(height uint32) {
-	s.LLeaderHeight = height //SetLeaderHeight()
+func (s *State) SetLLeaderHeight(height int) {
+	s.LLeaderHeight = uint32(height) //SetLeaderHeight()
 	s.LogPrintf("executeMsg", "Set LeaderHeight %d for %s", s.LLeaderHeight, atomic.WhereAmIString(1))
 }
